+++ dinero/tools/__init__.py
@@ -1,13 +1,30 @@
-from .calculate_vat import calculate_vat
-from .calculate_percentage import calculate_percentage
-from .calculate_simple_interest import calculate_simple_interest
-from .calculate_compound_interest import calculate_compound_interest
-from .calculate_markup import calculate_markup
+from .conversion import convert
+from .interest import calculate_compound_interest, calculate_simple_interest
+from .margin import (
+    calculate_cost_amount,
+    calculate_margin_portion,
+    calculate_selling_price,
+)
+from .markup import (
+    calculate_base_amount,
+    calculate_marked_up_amount,
+    calculate_markup_portion,
+)
+from .percentage import calculate_percentage
+from .vat import calculate_gross_amount, calculate_net_amount, calculate_vat_portion
 
 __all__ = [
-    "calculate_vat",
+    "calculate_gross_amount",
+    "calculate_net_amount",
+    "calculate_vat_portion",
     "calculate_percentage",
     "calculate_simple_interest",
     "calculate_compound_interest",
-    "calculate_markup",
+    "calculate_base_amount",
+    "calculate_marked_up_amount",
+    "calculate_markup_portion",
+    "calculate_cost_amount",
+    "calculate_margin_portion",
+    "calculate_selling_price",
+    "convert",
 ]