+++ tests/test_add.py
@@ -1,7 +1,7 @@
 import pytest
 
 from dinero import Dinero
-from dinero.currencies import USD, EUR
+from dinero.currencies import EUR, USD
 from dinero.exceptions import DifferentCurrencyError, InvalidOperationError
 
 
@@ -16,7 +16,7 @@ from dinero.exceptions import DifferentCurrencyError, InvalidOperationError
 def test_add_amount_str(amount, addend, total):
     assert amount + addend == total
     assert amount.add(addend) == total
-    assert amount.add(addend).equals_to(total)
+    assert amount.add(addend).eq(total)
 
 
 @pytest.mark.parametrize(
@@ -30,7 +30,7 @@ def test_add_amount_str(amount, addend, total):
 def test_add_amount_number(amount, addend, total):
     assert amount + addend == total
     assert amount.add(addend) == total
-    assert amount.add(addend).equals_to(total)
+    assert amount.add(addend).eq(total)
 
 
 @pytest.mark.parametrize(
@@ -54,7 +54,7 @@ def test_add_amount_number(amount, addend, total):
 def test_add_amount_mixed(amount, addend, total):
     assert amount + addend == total
     assert amount.add(addend) == total
-    assert amount.add(addend).equals_to(total)
+    assert amount.add(addend).eq(total)
 
 
 @pytest.mark.parametrize(
@@ -90,7 +90,7 @@ def test_sum_amount_mixed(amount, addend, total):
 )
 def test_different_currencies_error(amount, addend):
     with pytest.raises(DifferentCurrencyError):
-        amount + addend
+        amount + addend  # type: ignore
 
     with pytest.raises(DifferentCurrencyError):
         amount.add(addend)
@@ -106,7 +106,7 @@ def test_different_currencies_error(amount, addend):
 )
 def test_invalid_operation_error(amount, addend):
     with pytest.raises(InvalidOperationError):
-        amount + addend
+        amount + addend  # type: ignore
 
     with pytest.raises(InvalidOperationError):
         amount.add(addend)