+++ tests/test_operations.py
@@ -14,7 +14,7 @@ from dinero.exceptions import InvalidOperationError
 )
 def test_equal(obj_1, obj_2):
     assert obj_1 == obj_2
-    assert obj_1.equals_to(obj_2)
+    assert obj_1.eq(obj_2)
 
 
 @pytest.mark.parametrize(
@@ -26,7 +26,7 @@ def test_equal(obj_1, obj_2):
 )
 def test_not_equal(obj_1, obj_2):
     assert obj_1 != obj_2
-    assert obj_1.equals_to(obj_2) is False
+    assert obj_1.eq(obj_2) is False
 
 
 @pytest.mark.parametrize(
@@ -36,9 +36,9 @@ def test_not_equal(obj_1, obj_2):
         (Dinero(24.5, USD), Dinero(24.6, USD)),
     ],
 )
-def test_less_than(obj_1, obj_2):
+def test_lt(obj_1, obj_2):
     assert obj_1 < obj_2
-    assert obj_1.less_than(obj_2)
+    assert obj_1.lt(obj_2)
 
 
 @pytest.mark.parametrize(
@@ -49,9 +49,9 @@ def test_less_than(obj_1, obj_2):
         (Dinero(24.5, USD), Dinero(24.5, USD)),
     ],
 )
-def test_less_than_or_equal(obj_1, obj_2):
+def test_lte(obj_1, obj_2):
     assert obj_1 <= obj_2
-    assert obj_1.less_than_or_equal(obj_2)
+    assert obj_1.lte(obj_2)
 
 
 @pytest.mark.parametrize(
@@ -61,9 +61,9 @@ def test_less_than_or_equal(obj_1, obj_2):
         (Dinero(24.6, USD), Dinero(24.5, USD)),
     ],
 )
-def test_greater_than(obj_1, obj_2):
+def test_gt(obj_1, obj_2):
     assert obj_1 > obj_2
-    assert obj_1.greater_than(obj_2)
+    assert obj_1.gt(obj_2)
 
 
 @pytest.mark.parametrize(
@@ -74,9 +74,9 @@ def test_greater_than(obj_1, obj_2):
         (Dinero(24.5, USD), Dinero(24.5, USD)),
     ],
 )
-def test_greater_than_or_equal(obj_1, obj_2):
+def test_gte(obj_1, obj_2):
     assert obj_1 >= obj_2
-    assert obj_1.greater_than_or_equal(obj_2)
+    assert obj_1.gte(obj_2)
 
 
 @pytest.mark.parametrize(
@@ -89,31 +89,31 @@ def test_greater_than_or_equal(obj_1, obj_2):
 )
 def test_invalid_operation_error(amount, addend):
     with pytest.raises(InvalidOperationError):
-        amount == addend
+        amount == addend  # type: ignore
 
     with pytest.raises(InvalidOperationError):
-        amount.equals_to(addend)
+        amount.eq(addend)
 
     with pytest.raises(InvalidOperationError):
-        amount < addend
+        amount < addend  # type: ignore
 
     with pytest.raises(InvalidOperationError):
-        amount.less_than(addend)
+        amount.lt(addend)
 
     with pytest.raises(InvalidOperationError):
-        amount >= addend
+        amount >= addend  # type: ignore
 
     with pytest.raises(InvalidOperationError):
-        amount.less_than_or_equal(addend)
+        amount.lte(addend)
 
     with pytest.raises(InvalidOperationError):
-        amount > addend
+        amount > addend  # type: ignore
 
     with pytest.raises(InvalidOperationError):
-        amount.greater_than(addend)
+        amount.gt(addend)
 
     with pytest.raises(InvalidOperationError):
-        amount >= addend
+        amount >= addend  # type: ignore
 
     with pytest.raises(InvalidOperationError):
-        amount.greater_than_or_equal(addend)
+        amount.gte(addend)