+++ tests/test_tools.py
@@ -1,38 +1,13 @@
 import pytest
 
 from dinero import Dinero
-from dinero.currencies import USD, EUR, CLP, JPY
+from dinero.currencies import EUR, JPY, USD
 from dinero.exceptions import InvalidOperationError
 from dinero.tools import (
-    calculate_vat,
+    calculate_compound_interest,
     calculate_percentage,
     calculate_simple_interest,
-    calculate_compound_interest,
-    calculate_markup,
 )
-
-
-@pytest.mark.parametrize(
-    "amount, vat_rate, expected_vat_amount",
-    [
-        (Dinero(100, USD), 7.25, Dinero("6.76", USD)),
-        (Dinero(50, EUR), 21, Dinero("8.68", EUR)),
-        (Dinero(500, CLP), 19, Dinero("80", CLP)),
-        (100, 7.25, InvalidOperationError),
-        (Dinero(100, USD), "7.25", TypeError),
-        (Dinero(100, USD), -7.25, ValueError),
-    ],
-)
-def test_calculate_vat(amount, vat_rate, expected_vat_amount):
-    if isinstance(expected_vat_amount, type) and issubclass(
-        expected_vat_amount,
-        Exception,
-    ):
-        with pytest.raises(expected_vat_amount):
-            calculate_vat(amount, vat_rate)
-    else:
-        vat = calculate_vat(amount, vat_rate)
-        assert vat == expected_vat_amount
 
 
 @pytest.mark.parametrize(
@@ -119,24 +94,4 @@ def test_calculate_compound_interest(
             duration=duration,
             compound_frequency=compound_frequency,
         )
-        assert compound_interest.equals_to(expected)
-
-
-@pytest.mark.parametrize(
-    "cost, markup, expected_final_price, error",
-    [
-        (Dinero(1000, USD), 15, Dinero(1150, USD), None),
-        (Dinero(500, EUR), 10, Dinero(550, EUR), None),
-        (Dinero(2000, CLP), 20, Dinero(2400, CLP), None),
-        (1000, 15, None, InvalidOperationError),
-        (Dinero(1000, USD), "15", None, TypeError),
-        (Dinero(1000, USD), -15, None, ValueError),
-    ],
-)
-def test_calculate_markup(cost, markup, expected_final_price, error):
-    if error:
-        with pytest.raises(error):
-            calculate_markup(cost, markup)
-    else:
-        calculated_final_price = calculate_markup(cost, markup)
-        assert calculated_final_price == expected_final_price
+        assert compound_interest.eq(expected)